requirements.txt
<<<<<<< HEAD
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain-community
langchain_google_genai

=======
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain-community
langchain_google_genai

>>>>>>> 7f47261 (initial commit)
